requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.12
docopt==0.6.2
idna==3.3
IMAPClient==2.2.0
psycopg2-binary==2.9.3
requests==2.27.1
six==1.16.0
urllib3==1.26.8
yarg==0.1.9
